+++ Sources/Sentry/SentryOptions.m
@@ -21,7 +21,9 @@ SentryOptions ()
 
 NSString *const kSentryDefaultEnvironment = @"production";
 
-@implementation SentryOptions
+@implementation SentryOptions {
+    BOOL _enableTracingManual;
+}
 
 - (void)setMeasurement:(SentryMeasurementValue *)measurement
 {
@@ -75,6 +77,9 @@ NSString *const kSentryDefaultEnvironment = @"production";
         self.enableCaptureFailedRequests = YES;
         self.environment = kSentryDefaultEnvironment;
         self.enableTimeToFullDisplay = NO;
+
+        _enableTracing = NO;
+        _enableTracingManual = NO;
 #if SENTRY_HAS_UIKIT
         self.enableUIViewControllerTracing = YES;
         self.attachScreenshot = NO;
@@ -370,6 +375,10 @@ NSString *const kSentryDefaultEnvironment = @"production";
         self.tracesSampler = options[@"tracesSampler"];
     }
 
+    if ([options[@"enableTracing"] isKindOfClass:NSNumber.self]) {
+        self.enableTracing = [options[@"enableTracing"] boolValue];
+    }
+
     if ([options[@"inAppIncludes"] isKindOfClass:[NSArray class]]) {
         NSArray<NSString *> *inAppIncludes =
             [options[@"inAppIncludes"] filteredArrayUsingPredicate:isNSString];
@@ -470,14 +479,38 @@ NSString *const kSentryDefaultEnvironment = @"production";
     return [self isValidTracesSampleRate:sampleRate];
 }
 
+- (void)setEnableTracing:(BOOL)enableTracing
+{
+    //`enableTracing` is basically an alias to tracesSampleRate
+    // by enabling it we set tracesSampleRate to maximum
+    // if the user did not configured other ways to enable tracing
+    if ((_enableTracing = enableTracing)) {
+        if (_tracesSampleRate == nil && _tracesSampler == nil && _enableTracing) {
+            _tracesSampleRate = @1;
+        }
+    }
+    _enableTracingManual = YES;
+}
+
 - (void)setTracesSampleRate:(NSNumber *)tracesSampleRate
 {
     if (tracesSampleRate == nil) {
         _tracesSampleRate = nil;
     } else if ([self isValidTracesSampleRate:tracesSampleRate]) {
         _tracesSampleRate = tracesSampleRate;
+        if (!_enableTracingManual) {
+            _enableTracing = YES;
+        }
     } else {
         _tracesSampleRate = _defaultTracesSampleRate;
+    }
+}
+
+- (void)setTracesSampler:(SentryTracesSamplerCallback)tracesSampler
+{
+    _tracesSampler = tracesSampler;
+    if (_tracesSampler != nil && !_enableTracingManual) {
+        _enableTracing = YES;
     }
 }
 
@@ -489,8 +522,9 @@ NSString *const kSentryDefaultEnvironment = @"production";
 
 - (BOOL)isTracingEnabled
 {
-    return (_tracesSampleRate != nil && [_tracesSampleRate doubleValue] > 0)
-        || _tracesSampler != nil;
+    return _enableTracing
+        && ((_tracesSampleRate != nil && [_tracesSampleRate doubleValue] > 0)
+            || _tracesSampler != nil);
 }
 
 #if SENTRY_TARGET_PROFILING_SUPPORTED