+++ Tests/SentryTests/Protocol/SentryDebugMetaEquality.swift
@@ -4,8 +4,11 @@ extension DebugMeta {
     open override func isEqual(_ object: Any?) -> Bool {
         if  let other = object as? DebugMeta {
             return  uuid == other.uuid &&
+            debugID == other.debugID &&
                 type == other.type &&
-                name == other.name && imageSize == other.imageSize &&
+                name == other.name &&
+                codeFile == other.codeFile &&
+                imageSize == other.imageSize &&
                 imageAddress == other.imageAddress &&
                 imageVmAddress == other.imageVmAddress
         } else {